refactor: move test_utils under tests
