+++ healthcare/healthcare/report/diagnosis_trends/test_diagnosis_trends.py
@@ -7,7 +7,7 @@ from healthcare.healthcare.doctype.patient_appointment.test_patient_appointment 
 	create_practitioner,
 )
 from healthcare.healthcare.report.diagnosis_trends.diagnosis_trends import execute
-from healthcare.healthcare.test_utils import create_encounter
+from healthcare.tests.test_utils import create_encounter
 
 months = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"]
